weekly calendar - teste necessário

+++ src/pages/businessPage/Plans.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react'
 import { MainContainer, WrapContainer } from "../../theme";
-import { Text, Heading, Box,  } from "@chakra-ui/react";
+import { Text, Heading, Box, } from "@chakra-ui/react";
 import { DeleteIcon } from "@chakra-ui/icons";
-import { TextCard } from './TextCard';
-import { deleteItemById } from '../../api';
+import TextCard from './TextCard';
+import { deleteItemById, findAllItems } from '../../api';
 import { plansCol } from '../../api/config';
 import { PlanForm } from './PlanForm';
 import { DoubleClickText } from '../../components/DoubleClickText';
@@ -14,31 +14,24 @@ export const Plans = () => {
     const [loading, setLoading] = useState(false);
 
     useEffect(() => {
-        //buscar espaço
+        findAllItems(plansCol)
+            .then(res => setPlans(res))
+            .catch(err => console.log(err))
     }, []);
 
-    const mockPlans = [
-        { id: "1x-Mensal", monthlyPayment: "R$ 100,00" },
-        { id: "2x-Mensal", monthlyPayment: "R$ 100,00" },
-        { id: "3x-Mensal", monthlyPayment: "R$ 100,00" },
-        { id: "4x-Mensal", monthlyPayment: "R$ 100,00" },
-        { id: "5x-Mensal", monthlyPayment: "R$ 100,00" },
-        { id: "6x-Mensal", monthlyPayment: "R$ 100,00" }
-    ]
-
     const onDelete = (id) => {
-        setLoading(true)
-
-        deleteItemById(plansCol, id)
-            .catch(err => console.log(err))
-            .finally(setLoading(false))
-
+        if (window.confirm("Deletar este plano?")) {
+            setLoading(true)
+            deleteItemById(plansCol, id)
+                .catch(err => console.log(err))
+                .finally(setLoading(false))
+        }
     }
 
-    const list = mockPlans.map(plan => {
+    const list = plans?.length && plans.map(plan => {
         return (
             <TextCard
-                width={"150px"}
+                width={"180px"}
                 key={plan.id}
             >
                 <Box
@@ -48,12 +41,12 @@ export const Plans = () => {
                     gap={"0.1em"}
                     marginLeft={"0.5em"}
                 >
-                    <Text>{plan.id}</Text>
-                    <DoubleClickText text={plan.monthlyPayment} size={"10"} />
+                    <Text fontWeight={"bold"}>{plan.id}</Text>
+                    <DoubleClickText text={plan.price} size={"10"} />
                 </Box>
 
                 <DeleteIcon
-                    boxSize={"20px"}
+                    boxSize={"22px"}
                     onClick={() => onDelete(plan.id)}
                 />
             </TextCard>
@@ -64,7 +57,7 @@ export const Plans = () => {
     return (
         <MainContainer>
             <Heading size={"md"}>Planos</Heading>
-            <PlanForm />
+            <PlanForm loading={loading} setLoading={setLoading} />
             <br />
             <WrapContainer>
                 {list}
